+++ src/components/Navbar/Navbar.jsx
@@ -1,9 +1,35 @@
-import React from "react";
+//src/components/Navbar/Navbar.jsx
+import React, { useEffect, useState } from "react";
 import "./Navbar.css";
 import { NavLink } from "react-router-dom";
 import AuthStatus from "./AuthStatus";
+import LoginModal from "../Modal/LoginModal";
+import { loginUser } from "../../api/auth";
+import { useAuth } from "../../hooks/useAuth";
 
-function Navbar() {
+export default function Navbar() {
+  const { isLoggedIn } = useAuth();
+  const [showLoginModal, setShowLoginModal] = useState(false);
+
+  useEffect(() => {
+    if (!isLoggedIn) {
+      setShowLoginModal(true);
+    }
+  }, [isLoggedIn]);
+
+  const handleLogin = async (nickname, password) => {
+    try {
+      const data = await loginUser({ nickname, password });
+      localStorage.setItem("nickname", data.nickname);
+      localStorage.setItem("userId", data.id);
+      alert("로그인 성공!");
+      setShowLoginModal(false);
+      window.location.reload();
+    } catch (err) {
+      alert(err.message);
+    }
+  };
+
   return (
     <div className="navArea">
       <NavLink to="/" className="logo" />
@@ -22,8 +48,7 @@ function Navbar() {
         </NavLink>
       </div>
       <AuthStatus />
+      {showLoginModal && <LoginModal onLogin={handleLogin} />}
     </div>
   );
 }
-
-export default Navbar;
