+++ src/pages/LandingPage/LandingPage.jsx
@@ -1,26 +1,18 @@
 // src/pages/LandingPage/LandingPage.jsx
-import React from "react";
+import React, { useEffect, useState } from "react";
 import FetchTable from "../../components/FetchTable/FetchTable.jsx";
 import NoResult from "../../components/SearchBar/NoResult.jsx";
 import SearchBar from "../../components/SearchBar/SearchBar.jsx";
 import PageSizeSelector from "../../components/PageSizeSelector/PageSizeSelector";
-// import { invInitialData } from "../../config/invInitialData_v2";
-import { invInitialData } from "../../config/invInitialData_mock_80_with_description.js";
 import { useSearchFilter } from "../../hooks/useSearchFilter.js";
-import { useEffect, useState } from "react";
 import { usePagination } from "../../hooks/usePagination.js";
 import { usePageSize } from "../../hooks/usePageSize";
-
 import PaginationBtn from "../../components/DetailCompany/PaginationBtn.jsx";
 import { useCompanies } from "../../hooks/useCompanies.js";
 import SelectOption from "../../components/SelectOption/selectOption.jsx";
 import { LandingPageOptionsData } from "../../config/filterConfig.js";
-import { fetchFilteredData } from "../../api/api.jsx";
+// import { fetchFilteredData } from "../../api/api.jsx";
 import { fetchFilteredDataWJ } from "../../api/company.js";
-//로그인 작업 추가
-import LoginModal from "../../components/Modal/LoginModal.jsx";
-import { loginUser } from "../../api/auth";
-import { useAuth } from "../../hooks/useAuth";
 
 //나중에 config로 뺍시당
 const LandingPageColumns = [
@@ -34,8 +26,6 @@ const LandingPageColumns = [
 ];
 
 export default function LandingPage() {
-  const { isLoggedIn, nickname } = useAuth();
-  const [showLoginModal, setShowLoginModal] = useState(false);
   const [sortOption, setSortOption] = useState("totalInvestment_desc");
   const [sortBy, order] = sortOption.split("_");
   const { data, loading } = useCompanies({ sortBy, order });
@@ -66,29 +56,11 @@ export default function LandingPage() {
     };
     fetchData();
   }, [keyword, sortBy, order]);
-
-  useEffect(() => {
-    if (!isLoggedIn) {
-      setShowLoginModal(true);
-    }
-  }, []);
+  
 
   const handleCompanySortChange = (e) => {
     setSortOption(e.target.value);
     console.log(e.target.value);
-  };
-
-  const handleLogin = async (nickname, password) => {
-    try {
-      const data = await loginUser({ nickname, password });
-      localStorage.setItem("nickname", data.nickname);
-      localStorage.setItem("userId", data.id);
-      setShowLoginModal(false);
-      alert("로그인 성공!");
-      window.location.reload();
-    } catch (err) {
-      alert(err.message);
-    }
   };
 
   //현재 페이지의 데이터만 자르기 //요부분은 calculatePageIndex 함수로 따로 빼도될듯
@@ -130,8 +102,6 @@ export default function LandingPage() {
       ) : (
         <NoResult keyword={keyword} />
       )}
-
-      {showLoginModal && <LoginModal onLogin={handleLogin} />}
     </div>
   );
 }
